requirements.txt
fastapi==0.115.0
mypy==1.11.2
pydantic==2.9.2
python-dotenv==1.0.1
uvicorn==0.30.6

pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-testdox==3.1.0
pytest-cov==5.0.0
pytest-clarity==1.0.1
freezegun==1.5.1
httpx==0.27.2

alembic==1.13.3
flake8==7.1.1
flake8-unused-arguments==0.0.13
faker==30.8.0